+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+## Upcoming
+
+* [FIX] Webpack executables not found if dpat installed globally
+* [FEATURE] Add new build module which exports all the tools for building apps
+
 ## v0.3.4 - 2017-06-13
+
 * [FIX] Failure to resolve dpat root path 
 * [FIX] In dev mode, copy sdk distributions only if they have been added as dependencies